+++ CHANGELOG.md
@@ -4,9 +4,11 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+- Feature: `Carousel` passes props relating to visibility to its children.
+
 ## [2.77.0] - 2020-07-02
 
-Feature: 
+Feature:
 - Add `withoutMargin` prop on `Modal.Paragraph`.
 
 Fix: